requirements.txt
playwright
pyyaml
python-dotenv